moved prompter.js contents back into index
